refactor(login): move login handler to express
