+++ server/index.js
@@ -11,5 +11,6 @@ app.use(createWebSocketBridge('/api/bridge'))
 app.use(express.static(process.env.NODE_ENV === 'production' ? 'build' : 'public'))
 
 app.use('/api/session', require('./account'))
+app.use('/api/login', require('./login'))
 
 app.listen(port, () => console.log(`listening at http://localhost:${port}`))
